move modal actions file to appActions folder
